refactor(domain): replace old Navigator and NavigatorJgraphtAdapter with new domains
- fix broken tests
