--- src/main/java/wooteco/subway/domain/path/Navigator2.java
@@ -1,8 +1,0 @@
-package wooteco.subway.domain.path;
-
-import java.util.List;
-
-public interface Navigator2<V, E> {
-
-    List<E> calculateShortestPath(V source, V target);
-}
